Enable BDK async-interface and reqwest-default-tls features. Add isomorphic version of the log macro.
