+++ src/operations/bitcoin/balance.rs
@@ -2,7 +2,6 @@ use std::rc::Rc;
 
 use anyhow::Result;
 use bdk::{blockchain::esplora::EsploraBlockchain, database::MemoryDatabase, SyncOptions, Wallet};
-use bdk_macros::{maybe_async, maybe_await};
 
 use crate::data::{
     constants::{BITCOIN_EXPLORER_API, NETWORK},
@@ -28,7 +27,7 @@ pub async fn get_wallet(
         *NETWORK.read().unwrap(),
         MemoryDatabase::default(),
     )?;
-    maybe_await!(synchronize_wallet(&wallet))?;
+    synchronize_wallet(&wallet).await?;
     Ok(wallet)
 }
 
@@ -36,9 +35,8 @@ pub fn get_blockchain() -> EsploraBlockchain {
     EsploraBlockchain::new(&BITCOIN_EXPLORER_API.read().unwrap(), 100)
 }
 
-#[maybe_async]
-pub fn synchronize_wallet(wallet: &Wallet<MemoryDatabase>) -> Result<()> {
+pub async fn synchronize_wallet(wallet: &Wallet<MemoryDatabase>) -> Result<()> {
     let blockchain = get_blockchain();
-    maybe_await!(wallet.sync(&blockchain, SyncOptions::default()))?;
+    wallet.sync(&blockchain, SyncOptions::default()).await?;
     Ok(())
 }
