+++ src/operations/bitcoin/sign_psbt.rs
@@ -1,6 +1,5 @@
 use anyhow::{Error, Result};
 use bdk::{blockchain::Blockchain, database::MemoryDatabase, SignOptions, Wallet};
-use bdk_macros::maybe_await;
 use bitcoin::{consensus::serialize, util::psbt::PartiallySignedTransaction};
 use gloo_console::log;
 
@@ -16,7 +15,7 @@ pub async fn sign_psbt(
         let tx = psbt.extract_tx();
         log!("tx: {}", base64::encode(&serialize(&tx)));
         let blockchain = get_blockchain();
-        maybe_await!(blockchain.broadcast(&tx))?;
+        blockchain.broadcast(&tx).await?;
         Ok(())
     } else {
         Err(Error::msg(""))
