--- src/utils.rs
@@ -1,27 +1,0 @@
-use js_sys::Promise;
-use serde::de::DeserializeOwned;
-use wasm_bindgen::prelude::*;
-use wasm_bindgen_futures::JsFuture;
-
-pub fn set_panic_hook() {
-    // When the `console_error_panic_hook` feature is enabled, we can call the
-    // `set_panic_hook` function at least once during initialization, and then
-    // we will get better error messages if our code ever panics.
-    //
-    // For more details see
-    // https://github.com/rustwasm/console_error_panic_hook#readme
-    #[cfg(feature = "console_error_panic_hook")]
-    console_error_panic_hook::set_once();
-}
-
-pub async fn resolve(promise: Promise) -> JsValue {
-    JsFuture::from(promise).await.unwrap()
-}
-
-pub fn to_string(js_str: &JsValue) -> String {
-    js_str.as_string().unwrap()
-}
-
-pub fn json_parse<T: DeserializeOwned>(js_str: &JsValue) -> T {
-    serde_json::from_str(&js_str.as_string().unwrap()).expect("parsed json")
-}
