+++ tests/asset.rs
@@ -1,4 +1,4 @@
-use bdk_macros::await_or_block;
+use anyhow::Result;
 use bitmask_core::{get_vault, get_wallet_data, import_asset, save_mnemonic_seed};
 
 const MNEMONIC: &str =
@@ -9,8 +9,8 @@ const SEED_PASSWORD: &str = "";
 const ASSET: &str = "rgb1g2antx89ypjuat7jdth35d8xgqserckrhj9elkrhxhjhxch8sxqqguzmh6"; // BUX
 
 /// Test asset import
-#[test]
-fn asset_import() {
+#[tokio::test]
+async fn asset_import() -> Result<()> {
     // Import wallet
     let mnemonic_data = save_mnemonic_seed(
         MNEMONIC.to_owned(),
@@ -25,19 +25,21 @@ fn asset_import() {
     // Get vault properties
     let vault = get_vault(ENCRYPTION_PASSWORD.to_owned(), encrypted_descriptors).unwrap();
 
-    let asset = await_or_block!(import_asset(
+    let asset = import_asset(
         vault.rgb_tokens_descriptor.clone(),
         Some(ASSET.to_owned()),
         None,
         None,
-    ))
-    .unwrap();
+    )
+    .await?;
 
     assert_eq!(asset.id, ASSET);
 
     // Get wallet data
-    let wallet = await_or_block!(get_wallet_data(vault.rgb_tokens_descriptor, None)).unwrap();
+    let wallet = get_wallet_data(vault.rgb_tokens_descriptor, None).await?;
 
     // Parse wallet data
     assert_eq!(wallet.transactions, vec![]);
+
+    Ok(())
 }
